Move caching tests into a dedicated class
plus some tests refactorings
